[FEATURE] Add file embedding option to asset viewhelpers
The viewhelpers f:asset.css and f:asset.script are great but
missed an option to render referenced files inline.

A boolean option "inline" is now added to load the file contents
as inline styles or scripts. This is especially useful for
content elements which are used first in a page and need
some custom css to improve the Cumulative Layout Shift (CLS).

Resolves: #99510
Releases: main
Change-Id: Ic4282cd4a6ff00594a0aa0cbdf51f49d80806489
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/84424
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Simon Praetorius <simon@praetorius.me>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>