[TASK] Introduce TypolinkParameter object
This adds a new value object, containing
the resolved parameters of a typolink.

The object will be used as enriched value
for TCA type "link" properties in the
Record object. The enrichment will be added
with #103581.

Since this value is then passed to the view
instead of the plain typolink string, the
TypoLink ViewHelpers do also support the
new object.

Resolves: #104615
Related: #103581
Releases: main
Change-Id: Ifbac6d44ed05d5d793b951b38229891d6b219eb4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85619
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Jochen Roth <rothjochen@gmail.com>
Reviewed-by: Jochen Roth <rothjochen@gmail.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>