[BUGFIX] Prevent PHP error with empty link tag
If `<f:link.action />` is called with invalid parameters (e. g. a
non-existent page uid), the uri generation fails and returns an empty
string instead. In that case, only the provided tag content is used,
without any `<a>` tag. This fails with a PHP type error if no content
is provided and `$this->renderChildren()` returns `null`. That value
is now cast to a string to prevent the type error.

Resolves: #105141
Releases: main, 13.4, 12.4
Change-Id: Iae87d14d15eef0199b05627021ed58924484a431
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87091
Tested-by: Simon Praetorius <simon@praetorius.me>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Simon Praetorius <simon@praetorius.me>