[BUGFIX] Ensure AVIF files are only written when format is supported
The implementation with #105704 contained a bug that did not properly
check if 'avif' writing support was enabled, and did not trigger
the proper fallback chain.

This is moved now so that an 'avif' target file format is checked
in both cases:

a) For generic 'web' format, keeping an existing AVIF file and
   resizing it. Note this fallback is currently only used in
   legacy cases.

b) For the specific 'avif' output format.

Tests are added to ensure this works.

A coupled bug #105847 is fixed to ensure that when the
conversion of a format fails, the actually used file extension
will be persisted to the processed file. This scenario only really
happens with the current AVIF implementation, which is why
this bugfix is not made independently.

Without this bugfix, when a file like "test.jpg" should
be converted to "test.avif", but no AVIF processing is
available (GraphicsMagick), the resulting file would be
stored as "csm_test_4815162342.avif" - containing
JPEG contents though. With this fix, the file will be
persisted as "csm_test_4815162342.avif.jpg" to also
indicate a processing fallback chain took place.

Resolves: #105846
Resolves: #105847
Related: #105704
Related: #105847
Releases: main, 13.4
Change-Id: If2c0884a8e1de124f5ce2e6e6de22c8ec41bbb26
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87625
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>