[TASK] Update Fluid Standalone to 4.3.0
This patch updates Fluid to version 4.3.0. This release includes the
components feature, which was funded as a community budget idea.
As the version number states, the feature didn't introduce any
breaking changes to Fluid.

With this release, a new API has been introduced for ViewHelper
resolving: The `ViewHelperResolverDelegateInterface`. Since
Fluid creates these objects from a FQCN (just like ViewHelpers),
it makes sense to also support TYPO3's dependency injection
there. For this to work, the factory method in Fluid's
`ViewHelperResolver` is overridden. This is covered by
a functional test and an example extension.

Release notes on GitHub:
https://github.com/TYPO3/Fluid/releases/tag/4.3.0

Used commands:

> composer req typo3fluid/fluid:^4.3.0
> composer req -d typo3/sysext/adminpanel --no-update \
    typo3fluid/fluid:^4.3.0
> composer req -d typo3/sysext/core --no-update \
    typo3fluid/fluid:^4.3.0
> composer req -d typo3/sysext/fluid --no-update \
    typo3fluid/fluid:^4.3.0
> composer req -d typo3/sysext/redirects --no-update \
    typo3fluid/fluid:^4.3.0

Resolves: #106951
Releases: main, 13.4
Change-Id: I81cc6f7817a18f410a75a1e850b4625885f65682
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89844
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Simon Praetorius <simon@praetorius.me>