[TASK] Auto-configure Fluid ViewHelperResolver delegates
With Fluid 4.3, the `ViewHelperResolverDelegateInterface` was
introduced, which allows to "take over" a full ViewHelper
namespace instead of just one ViewHelper instance. When Fluid
was updated to 4.3 in the Core, dependency injection was added
to ViewHelperResolver delegates. For this to work properly,
these services need to be defined as `public` since Fluid needs
to access them on-demand based on the namespaces defined
in the template.

This patch extends the existing auto-configuration for ViewHelpers
to also auto-configure ViewHelperResolver delegates as `public`
based on the implemented interface.

Resolves: #107207
Related: #106951
Releases: main, 13.4
Change-Id: I28ecad6f77cc3c29ac5e0030b00f459e3823c899
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90256
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>