[TASK] Simplify <f:render> debugging implementation
In an effort to reduce custom code in ViewHelper resolving, the
`<f:render>` debugging implementation of the admin panel is moved to
EXT:adminpanel. The activation logic, which depends on request
information, is now executed with each ViewHelper call (where this
information is available through proper APIs), but only if
EXT:adminpanel is installed.

This makes the conditional namespace merging in the
`ViewHelperResolver` obsolete and allows us to deprecate the custom
"debuggable" `RenderViewHelper` implementation in EXT:fluid.

Some backend functional tests need to be adjusted to provide
a proper backend request to (eventually) their Fluid template.
In addition, some outdated code comments are removed or
moved to the appropriate position.

Resolves: #107208
Releases: main
Change-Id: I09a13d66ff108e5d49aceb0b91482b569272c0c0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90257
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Matthias Vogel <typo3@kanti.de>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Tested-by: Matthias Vogel <typo3@kanti.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>