[TASK] Avoid unintentional null value for form ViewHelpers
With Fluid 5, `null` being passed to `TagBuilder` will behave
differently: Before, this lead to `value=""`, now it will lead to
the attribute being omitted from the resulting HTML tag.

For Form ViewHelpers this can be problematic because HTML form elements
sometimes behave differently: The `<option>` tag, for example, falls
back to the option's label if the `value` attribute is omitted, which
is not what we usually want (and it differs from the current
implementation).

In preparation for the raise of Fluid to 5.0.0, the affected form
ViewHelper are adjusted to always cast the value to string, which
makes sure that their behavior stays the same after the raise.

Resolves: #108152
Related: #107660
Related: #108148
Releases: main
Change-Id: I3626b34c4a3a50987b073090f6eaea53d9286df2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91696
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>