[BUGFIX] Allow link-sections in *.action viewhelpers to be pure numbers
Albeit HTML id-attribute values may not be a pure number,
there seem to be usages for this in the wild.

Since using pure numbers was possible in the past,
we allow this again for the time being.

Note: While the Fluid attribute is already typed 'string',
we still receive an int for this attribute if its value is only
numbers. This is changed with Fluid v5, hence this fix
is only targeting v13.

Resolves: #108338
Releases: 13.4
Change-Id: I18270d99ce4dc033c458a559a0cc5ec0a8e50215
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91917
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>